layout refactor
