+++ src/pages/student/StudentRegister.jsx
@@ -1,12 +1,14 @@
 import fakeLogo from "/Ellipse.png";
 import fakeProfile from "/profile-pic.png";
-import TeacherForm from "../components/TeacherForm";
+import StudentForm from "../components/StudentForm";
+import Layout from "../components/Layout";
+
 import { useAuth } from "@arcana/auth-react";
 
-export default function TeacherRegister() {
+export default function StudentRegister() {
   const { user } = useAuth();
   return (
-    <div className="w-screen h-screen bg-gradient-to-br from-white to-rose-100 flex flex-col justify-between items-center">
+    <Layout>
       <div className="mb-auto mt-10 px-4 w-full flex justify-between items-center">
         <div className="flex justify-start items-center">
           <img className=" w-14 h-14" src={fakeLogo} alt="logo" />
@@ -19,10 +21,7 @@ export default function TeacherRegister() {
         />
       </div>
       <h2 className="text-2xl font-semibold">Add you details</h2>
-      <TeacherForm name={user.name} />
-      <p className={"mt-auto text-gray-400"}>
-        ©Splash Screen | All rights reserved
-      </p>
-    </div>
+      <StudentForm name={user.name} />
+    </Layout>
   );
 }
