+++ src/pages/teacher/TeacherLogin.jsx
@@ -1,5 +1,7 @@
 import googleIcon from "/google-icon.jpeg";
 import fakeLogo from "/Ellipse.png";
+import Layout from "../components/Layout";
+
 import { useAuth, Auth } from "@arcana/auth-react";
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
@@ -19,11 +21,9 @@ function TeacherLogin() {
     }
   }, [user, isLoggedIn]);
 
-  user = { adderss, email, id, name, picture, publicKey };
-
   return (
-    <div className="w-screen h-screen bg-gradient-to-br from-white to-rose-100 flex flex-col justify-center items-center gap-6">
-      <img className="mt-10 h-20 w-20" src={fakeLogo} alt="logo" />
+    <Layout>
+      <img className="my-10 h-20 w-20" src={fakeLogo} alt="logo" />
       {/* <div
         onClick={handleLogin}
         className="my-auto px-1 rounded-md flex justify-start items-center gap-6 h-14 w-3/5 max-w-md bg-google-blue hover:bg-google-blue-active cursor-pointer"
@@ -38,10 +38,7 @@ function TeacherLogin() {
         <p className=" text-white">Sign in with google</p>
       </div> */}
       <Auth externalWallet={true} theme="light" />
-      <p className={"mt-auto text-gray-400 "}>
-        ©Splash Screen | All rights reserved
-      </p>
-    </div>
+    </Layout>
   );
 }
